Create a reusable sortable table component
