+++ web/src/features/fireWeather/components/NoonWeatherValueTable.tsx
@@ -1,58 +1,32 @@
-import React, { useState } from 'react'
-import { makeStyles } from '@material-ui/core/styles'
-import Table from '@material-ui/core/Table'
-import TableBody from '@material-ui/core/TableBody'
-import TableCell from '@material-ui/core/TableCell'
-import TableContainer from '@material-ui/core/TableContainer'
-import TableRow from '@material-ui/core/TableRow'
-import TableHead from '@material-ui/core/TableHead'
-import TableSortLabel from '@material-ui/core/TableSortLabel'
-import Paper from '@material-ui/core/Paper'
-import Typography from '@material-ui/core/Typography'
+import React from 'react'
 
 import { ModelValue } from 'api/modelAPI'
 import { NoonForecastValue } from 'api/forecastAPI'
 import { MODEL_VALUE_DECIMAL } from 'utils/constants'
-import { getDatetimeComparator, Order } from 'utils/table'
+import SortableTableByDatetime, {
+  Column
+} from 'features/fireWeather/components/SortableTableByDatetime'
 
 /**
  * Reusable component used to display noon forecasts (issued by forecasters)
  * and/or model forecasts (generated by computer-run models) in table format
  */
 
-const useStyles = makeStyles({
-  display: {
-    paddingTop: 8,
-    paddingBottom: 8
-  },
-  paper: {
-    width: '100%'
-  },
-  tableContainer: {
-    maxHeight: 300
-  },
-  title: {
-    paddingBottom: 4
-  }
-})
-
-interface Column {
-  id:
-    | 'datetime'
-    | 'temperature'
-    | 'relative_humidity'
-    | 'wind_direction'
-    | 'wind_speed'
-    | 'total_precipitation'
-  label: string
-  minWidth?: number
-  maxWidth?: number
-  align?: 'left' | 'right' | 'center'
-  format?: (value: number) => string | number
-}
+type Id =
+  | 'datetime'
+  | 'temperature'
+  | 'relative_humidity'
+  | 'wind_direction'
+  | 'wind_speed'
+  | 'total_precipitation'
 
 const columns: Column[] = [
-  { id: 'datetime', label: 'Date (PDT)', align: 'left' },
+  {
+    id: 'datetime',
+    label: 'Date (PDT)',
+    align: 'left',
+    formatDt: (value: string) => value.slice(0, 10)
+  },
   {
     id: 'temperature',
     label: 'Temp (°C)',
@@ -90,89 +64,13 @@ const columns: Column[] = [
 ]
 
 interface Props {
-  rows: (ModelValue | NoonForecastValue)[] | undefined
-  title: string
   testId: string
+  title: string
+  rows: (ModelValue | NoonForecastValue)[] | undefined
 }
 
 const NoonWeatherValueTable = (props: Props) => {
-  const classes = useStyles()
-  const [order, setOrder] = useState<Order>('asc')
-
-  if (!props.rows) {
-    return null
-  }
-
-  const rowsSortedByDatetime = [...props.rows].sort(getDatetimeComparator(order))
-  const toggleDatetimeOrder = () => {
-    setOrder(order === 'asc' ? 'desc' : 'asc')
-  }
-
-  return (
-    <div className={classes.display} data-testid={props.testId}>
-      <Typography className={classes.title} variant="subtitle2" component="div">
-        {props.title}
-      </Typography>
-      <Paper className={classes.paper} elevation={1}>
-        <TableContainer className={classes.tableContainer}>
-          <Table stickyHeader size="small" aria-label="noon weather data table">
-            <TableHead>
-              <TableRow>
-                {columns.map(column => {
-                  const canSort = column.id === 'datetime'
-
-                  return (
-                    <TableCell
-                      key={column.id}
-                      align={column.align}
-                      style={{ minWidth: column.minWidth, maxWidth: column.maxWidth }}
-                      sortDirection={canSort ? order : false}
-                    >
-                      {canSort ? (
-                        <TableSortLabel
-                          active={canSort}
-                          direction={order}
-                          onClick={toggleDatetimeOrder}
-                        >
-                          {column.label}
-                        </TableSortLabel>
-                      ) : (
-                        column.label
-                      )}
-                    </TableCell>
-                  )
-                })}
-              </TableRow>
-            </TableHead>
-
-            <TableBody>
-              {rowsSortedByDatetime.map((row, idx) => (
-                <TableRow key={idx} hover tabIndex={-1}>
-                  {columns.map(column => {
-                    const value = row[column.id]
-                    let display = null
-
-                    if (typeof value === 'string' && column.id === 'datetime') {
-                      display = value.slice(0, 10)
-                    }
-                    if (typeof value === 'number' && column.format) {
-                      display = column.format(value)
-                    }
-
-                    return (
-                      <TableCell key={column.id} align={column.align}>
-                        {display}
-                      </TableCell>
-                    )
-                  })}
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </TableContainer>
-      </Paper>
-    </div>
-  )
+  return <SortableTableByDatetime {...props} columns={columns} />
 }
 
 export default React.memo(NoonWeatherValueTable)
