+++ web/src/features/fireWeather/components/SortableTableByDatetime.tsx
@@ -1,0 +1,141 @@
+import React, { useState } from 'react'
+import { makeStyles } from '@material-ui/core/styles'
+import Table from '@material-ui/core/Table'
+import TableBody from '@material-ui/core/TableBody'
+import TableCell from '@material-ui/core/TableCell'
+import TableContainer from '@material-ui/core/TableContainer'
+import TableRow from '@material-ui/core/TableRow'
+import TableHead from '@material-ui/core/TableHead'
+import TableSortLabel from '@material-ui/core/TableSortLabel'
+import Paper from '@material-ui/core/Paper'
+import Typography from '@material-ui/core/Typography'
+
+import { getDatetimeComparator, Order } from 'utils/table'
+
+const useStyles = makeStyles({
+  display: {
+    paddingBottom: 12
+  },
+  paper: {
+    width: '100%'
+  },
+  tableContainer: {
+    maxHeight: 280
+  },
+  title: {
+    paddingBottom: 4
+  }
+})
+
+interface WeatherValue {
+  datetime: string
+  temperature?: number | null
+  relative_humidity?: number | null
+  wind_direction?: number | null
+  wind_speed?: number | null
+  precipitation?: number | null
+  total_precipitation?: number | null
+  ffmc?: number | null
+  isi?: number | null
+  fwi?: number | null
+}
+
+export interface Column {
+  id: keyof WeatherValue
+  label: string
+  minWidth?: number
+  maxWidth?: number
+  align?: 'left' | 'right' | 'center'
+  format?: (value: number) => string | number
+  formatDt?: (dt: string) => string
+}
+
+interface Props<R> {
+  testId: string
+  title: string
+  rows: R[] | undefined
+  columns: Column[]
+}
+
+function SortableTableByDatetime<R extends WeatherValue>(props: Props<R>) {
+  const classes = useStyles()
+  const [order, setOrder] = useState<Order>('asc')
+
+  if (!props.rows) {
+    return null
+  }
+
+  const rowsSortedByDatetime = [...props.rows].sort(getDatetimeComparator(order))
+  const toggleDatetimeOrder = () => {
+    setOrder(order === 'asc' ? 'desc' : 'asc')
+  }
+
+  return (
+    <div className={classes.display} data-testid={props.testId}>
+      <Typography className={classes.title} component="div" variant="subtitle2">
+        {props.title}
+      </Typography>
+
+      <Paper className={classes.paper} elevation={1}>
+        <TableContainer className={classes.tableContainer}>
+          <Table stickyHeader size="small" aria-label="sortable wx table">
+            <TableHead>
+              <TableRow>
+                {props.columns.map(column => {
+                  const canSort = column.id === 'datetime'
+
+                  return (
+                    <TableCell
+                      key={column.id}
+                      align={column.align}
+                      style={{ minWidth: column.minWidth, maxWidth: column.maxWidth }}
+                      sortDirection={canSort ? order : false}
+                    >
+                      {canSort ? (
+                        <TableSortLabel
+                          active={canSort}
+                          direction={order}
+                          onClick={toggleDatetimeOrder}
+                        >
+                          {column.label}
+                        </TableSortLabel>
+                      ) : (
+                        column.label
+                      )}
+                    </TableCell>
+                  )
+                })}
+              </TableRow>
+            </TableHead>
+
+            <TableBody>
+              {rowsSortedByDatetime.map((row, idx) => (
+                <TableRow key={idx} hover tabIndex={-1}>
+                  {props.columns.map(column => {
+                    const value = row[column.id]
+                    let display = null
+
+                    if (typeof value === 'string' && column.formatDt) {
+                      display = column.formatDt(value)
+                    }
+                    if (typeof value === 'number' && column.format) {
+                      display = column.format(value)
+                    }
+
+                    return (
+                      <TableCell key={column.id} align={column.align}>
+                        {display}
+                      </TableCell>
+                    )
+                  })}
+                </TableRow>
+              ))}
+            </TableBody>
+          </Table>
+        </TableContainer>
+      </Paper>
+    </div>
+  )
+}
+
+export default React.memo(SortableTableByDatetime)
